requirements.txt
python-docx==1.1.0
pdfplumber==0.11.4
streamlit==1.35.0
pydantic==2.8.2
typing-extensions
# difflib (stdlib, listed for completeness)
